+++ src/main/kotlin/com/gemini/krakenbot/repository/impl/SqliteTradeRepositoryImpl.kt
@@ -6,14 +6,13 @@ import com.gemini.krakenbot.model.TradeRecord
 import com.gemini.krakenbot.repository.TradeRepository
 import com.gemini.krakenbot.repository.TradeSummaryStats
 import com.gemini.krakenbot.repository.table.*
+import com.gemini.krakenbot.service.isWithinRelativeTolerance
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.inList
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.less
-import org.jetbrains.exposed.sql.transactions.transaction
 import org.jetbrains.exposed.sql.statements.UpdateBuilder
-import com.gemini.krakenbot.service.isWithinRelativeTolerance
+import org.jetbrains.exposed.sql.transactions.transaction
 import org.slf4j.LoggerFactory
-import java.io.IOException
 import java.math.BigDecimal
 import java.math.RoundingMode
 import java.time.Instant